extract program launch scheduling policy
